LayoutTests/platform/chromium/fast/forms/calendar-picker/calendar-picker-datetime.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=datetime id=datetime value="2011-12-31T12:53:45Z"> 
<script>
description("Tests if calendar picker works for datetime as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('datetime'), test);

function test() {
    shouldBeEqualToString('highlightedValue()', '2011-12-31');
    shouldBeEqualToString('selectedValue()', '2011-12-31');

    eventSender.keyDown('rightArrow');
    shouldBeEqualToString('highlightedValue()', '2012-01-01');
    shouldBeEqualToString('selectedValue()', '2011-12-31');

    eventSender.keyDown('\n');
    shouldBeEqualToString('highlightedValue()', '2012-01-01');
    shouldBeEqualToString('selectedValue()', '2012-01-01');
    shouldBeEqualToString('document.getElementById("datetime").value', '2012-01-01T12:53:45Z');

    finishJSTest();
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
